DotCloud Build File
The DotCloud Build File, ``dotcloud.yml``, describes our stack.

We declare a single "ruby" service. We use "ruby" because it features
two useful things for us:

* a Ruby interpreter (obviously!) and a mechanism to install gems;
* a public-facing HTTP server (nginx).

The role and syntax of the DotCloud Build File is explained in further
detail in the documentation, at http://docs.dotcloud.com/guides/build-file/.