Postinstall Script
To start god (and, therefore, Thin) automatically, we will use a
``postinstall`` script.

We could use SSH and start god manually each time we push the service,
but that would not be very convenient. To make sure that our processes
are started automatically when the service is pushed or scaled, we will
write a short postinstall script. The postinstall script is executed
automatically after each push, and after each deployment of a new
scaled instance.

In this script, we just have to invoke ``god`` and give him the
configuration file name.

Remember to make sure that the ``postinstall`` script is executable,
with e.g. ``chmod +x postinstall``.