Install God Process Manager
We want something to start, stop, and monitor the "thin" process
better than what we could do with crude scripts. We will use the
"god" process manager. It is the Ruby equivalent of e.g. Supervisor
or Monit.

To tell DotCloud to install it automatically for us, we simply
add it to our Gemfile.